events: Factor out the xkb keysym to scancode conversion from the X11 driver
